+++ CHANGELOG.md
@@ -7,6 +7,7 @@
 * [added] `unmute` command in AppleScript
 * [added] `playback state` read/write attribute in AppleScript
 * [fixed] A few memory leaks have been resolved
+* [fixed] Use Pandora's `sync` API for real
 
 # Version 1.0.13 (11/20/11)